refactor and tests
