+++ src/index.ts
@@ -1,7 +1,8 @@
 export { ConnectionConfig, ConnectionConfigProps } from './ConnectionConfig';
 export { ConfigSelect, InlineInput } from './sql/ConfigEditor';
 export { ResourceSelector, ResourceSelectorProps } from './sql/ResourceSelector';
-export { QueryCodeEditor, QuerySelect, FormatSelect, FillValueSelect } from './sql/QueryEditor';
+export { SQLQuery } from './sql/types';
+export { QueryCodeEditor, FormatSelect, FillValueSelect } from './sql/QueryEditor';
 export * from './types';
 export * from './regions';
 export * from './providers';
