+++ src/sql/QueryEditor/FillValueSelect.tsx
@@ -44,6 +44,7 @@ export function FillValueSelect<TQuery extends DataQuery & Record<string, any>>(
             });
             props.onRunQuery();
           }}
+          className="width-12"
         />
       </InlineField>
       {props.query.fillMode?.mode === FillValueOptions.Value && (
