+++ src/sql/QueryEditor/__mocks__/query.ts
@@ -1,15 +1,8 @@
-import { DataQuery } from '@grafana/data';
-import { FillValueOptions } from '../FillValueSelect';
+import { SQLQuery } from '../../types';
 
 export enum SQLOptions {
   Table,
   TimeSeries,
-}
-
-export interface SQLQuery extends DataQuery {
-  rawSQL: string;
-  format?: SQLOptions;
-  fillMode?: { mode: FillValueOptions; value?: number };
 }
 
 export const mockQuery: SQLQuery = { rawSQL: 'select * from table', refId: '', fillMode: { mode: 0 }, format: 0 };
