+++ src/sql/QueryEditor/index.ts
@@ -1,4 +1,3 @@
 export { QueryCodeEditor } from './QueryCodeEditor';
-export { QuerySelect } from './QuerySelect';
 export { FormatSelect } from './FormatSelect';
 export { FillValueSelect } from './FillValueSelect';
